+++ src/router.js
@@ -4,7 +4,6 @@ import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import RankingScreen from './screens/RankingScreen';
 import InfoScreen from './screens/InfoScreen';
-import TrialScreen from './screens/trialScreen';
 import SignInScreen from './screens/SignInScreen';
 import HomeScreen from './screens/HomeScreen';
 import firebaseConfig from './firebase';
@@ -42,13 +41,6 @@ const Routes = () => (
       <Stack.Screen
         name="InfoScreen"
         component={InfoScreen}
-        options={{
-          headerShown: false,
-        }}
-      />
-      <Stack.Screen
-        name="TrialScreen"
-        component={TrialScreen}
         options={{
           headerShown: false,
         }}
