+++ src/styles/InfoScreenStyle.js
@@ -42,85 +42,90 @@ const styles = StyleSheet.create({
     fontSize: 18,
     fontWeight: 'bold',
   },
+  divider: { backgroundColor: theme.colors.ourgreen, height: 6 },
   container: {
-    // height: screenHeight - 130,
-    justifyContent: 'space-evenly',
-    width: screenWidth,
+    flex: 1,
+    flexDirection: 'row',
+    backgroundColor: theme.colors.secondary,
+  },
+  leftcontainer: {
+    backgroundColor: theme.colors.white,
     flex: 1,
     flexDirection: 'column',
+    width: screenWidth / 2,
+    height: 'auto',
   },
-  divider: { backgroundColor: theme.colors.ourgreen, height: 6 },
-  info_view1: {
+  rightcontainer: {
+    backgroundColor: theme.colors.white,
+    flex: 1,
     flexDirection: 'row',
-    marginTop: 25,
-    width: screenWidth,
+    flexWrap: 'wrap',
+  },
+  leftheadertext: {
+    fontSize: 20,
+    marginLeft: 20,
+    padding: 24,
+    width: 170,
+  },
+  moveright: {
+    marginLeft: screenWidth - 260,
+    padding: 20,
+    flexDirection: 'row',
+  },
+  iconmoveright: {
+    padding: 20,
+    flexDirection: 'row',
+  },
+  nBtn: {
+    width: 30,
+    borderWidth: 1,
+    backgroundColor: theme.colors.white,
+    borderRadius: 10,
+    borderColor: theme.colors.button_border,
+    height: 30,
+  },
+  nBtnText: {
+    padding: 5,
+    alignSelf: 'center',
+  },
+  nBtn2: {
+    width: 80,
+    borderWidth: 1,
+    backgroundColor: theme.colors.white,
+    borderRadius: 10,
+    borderColor: theme.colors.button_border,
+    height: 40,
+  },
+  nBtnText2: {
+    padding: 10,
+    alignSelf: 'center',
   },
   card: {
     width: 80,
     backgroundColor: theme.colors.white,
     borderColor: theme.colors.button_border,
-    borderWidth: 2,
+    height: 60,
+    marginLeft: 20,
+    borderWidth: 1,
   },
   carddivider: {
     color: theme.colors.button_border,
-    height: 3,
+    height: 2,
   },
-  info_text: {
-    fontSize: 20,
-    marginLeft: 20,
+  button: {
+    backgroundColor: theme.colors.secondary,
+    height: 50,
+    justifyContent: 'center',
   },
-  info_view2: {
-    flexDirection: 'row',
-    marginTop: 40,
-    width: screenWidth,
-  },
-  info_view3: {
-    flexDirection: 'row',
-    marginTop: 25,
-    width: screenWidth,
-  },
-  info_view4: {
-    flexDirection: 'row',
-    marginTop: 25,
-    width: screenWidth,
-  },
-  info_view5: {
-    flexDirection: 'row',
-    marginTop: 25,
-    width: screenWidth,
-  },
-  nBtn: {
-    width: 30,
-    elevation: 8,
-    backgroundColor: theme.colors.white,
-    borderRadius: 10,
-    borderWidth: 2,
-    borderColor: theme.colors.button_border,
-    marginLeft: screenWidth - 350,
-    height: 30,
-  },
-  nBtn2: {
-    width: 80,
-    elevation: 8,
-    backgroundColor: theme.colors.white,
-    borderRadius: 10,
-    borderWidth: 2,
-    borderColor: theme.colors.button_border,
-    marginLeft: screenWidth - 350,
-    height: 40,
-  },
-  nBtnText2: {
-    paddingTop: 10,
-    alignSelf: 'center',
-  },
-  nBtnText: {
-    paddingTop: 5,
-    alignSelf: 'center',
+  buttontext: {
+    fontSize: 15,
+    fontWeight: 'bold',
+    justifyContent: 'center',
   },
   btn_view_info: {
     flexDirection: 'row',
     marginLeft: 10,
-    marginTop: 30,
+    marginBottom: 20,
     justifyContent: 'space-evenly',
   },
   btn_change: {
@@ -135,21 +140,6 @@ const styles = StyleSheet.create({
     backgroundColor: theme.colors.ourgreen,
     borderRadius: 10,
   },
-  button: {
-    backgroundColor: theme.colors.secondary,
-    borderRadius: 20,
-    height: 40,
-    justifyContent: 'center',
-    marginTop: 20,
-    marginVertical: 10,
-    marginLeft: 90,
-    width: 230,
-  },
-  buttontext: {
-    fontSize: 15,
-    fontWeight: 'bold',
-    justifyContent: 'center',
-  },
   roundButton: {
     marginVertical: 5,
     width: 30,
@@ -158,7 +148,7 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     borderRadius: 100,
-    borderWidth: 2,
+    borderWidth: 1,
     borderColor: theme.colors.button_border,
     backgroundColor: theme.colors.white,
   },
@@ -170,7 +160,7 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     borderRadius: 100,
-    borderWidth: 2,
+    borderWidth: 1,
     borderColor: theme.colors.button_border,
     backgroundColor: 'blue',
     shadowColor: theme.colors.white,
